flex CSS, data-wiki=lineup, insertAdjacentHTML instead of innerHTML
re-learned an old DOM trick in insertAdjacentHTML() which is similar
to appendChild() except that accepts a string of HTML instead of a
Node.

CSS changes get the footer against the bottom of the page for
firefox.

data-wiki=lineup anticipates some changes that might allow view.js to
insert a read-only wiki into a section of a page instead of directly
into the body.  (this part is incomplete, but committing now to take
before making this an even bigger change)