load types statically
Not completely convinced this is the right thing. However, this is a
static and read-only wiki. We should therefore know at publish time
all the types our site will support.

In any case, the code seems a good deal more straightforward this
way. We will trust that feeling for now.